+++ test/functional/groups.js
@@ -84,7 +84,7 @@ describe('GroupsController', () => {
         })
         .end((err) => {
           err.should.not.be.empty;
-          err.status.should.eql(500);
+          err.status.should.eql(422);
           done();
         });
     });
@@ -100,7 +100,7 @@ describe('GroupsController', () => {
         })
         .end((err) => {
           err.should.not.be.empty;
-          err.status.should.eql(500);
+          err.status.should.eql(422);
           err.response.error.should.have.property('text');
           JSON.parse(err.response.error.text).err.should.eql('Invalid username');
           done();
@@ -118,7 +118,7 @@ describe('GroupsController', () => {
         })
         .end((err) => {
           err.should.not.be.empty;
-          err.status.should.eql(500);
+          err.status.should.eql(422);
           err.response.error.should.have.property('text');
           JSON.parse(err.response.error.text).err.should.eql('Invalid username');
           done();
@@ -155,8 +155,7 @@ describe('GroupsController', () => {
             .get(`${app.context.config.host}/v1/users/Luna/subscriptions`)
             .query({ authToken: context.authToken })
             .end((err, res) => {
-              res.status.should.not.eql(404);
-              res.status.should.not.eql(500);
+              res.status.should.eql(200);
               res.body.should.not.be.empty;
               res.body.should.have.property('subscribers');
               res.body.should.have.property('subscriptions');